### 2025-08-13 #8793 @editorialbot
[PRE REVIEW]: vkCompViz: Universal C++ Library for GPU-Based Experiments
**Submitting author:** @ichlubna (<a href="http://orcid.org/0000-0003-3126-0545">Tomáš Chlubna</a>)
**Repository:** https://github.com/ichlubna/vkCompViz
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
