### 2025-08-13 #8793 @danielskatz
👋 @ichlubna - Given the relatively short commit history for your software, I'm going to ask the editors to discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or two.

Also, I think you can remove the line about equal author contributions in your paper, since there is just one author. After you do so, you can use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
