### 2025-08-13 188f1f4
Statement fix and doi url duplicate
