+++ paper/paper.bib
@@ -132,7 +132,6 @@ address = {New York, NY, USA},
 volume = {48},
 number = {4},
 issn = {0360-0300},
-url = {https://doi.org/10.1145/2873053},
 doi = {10.1145/2873053},
 month = mar,
 articleno = {60},
