### 2025-08-13 #8793 @ichlubna
Thank you @danielskatz for the info. I removed the equal contributions statement. I am looking forward to the editorial evaluation :-). Please note that some commits are quite long and Vulkan development is full of caveats that my software aims to eliminate for the use case of quick scientific experiments.

In the meantime, according to the initial instructions, I went through the list of reviewers and found that the area of expertise of some of them (GPU programming and graphics) aligns with the topic of the submission, the usernames are:
jdoerfert
JPenuchot
thivinanandh
dhelmrich

I would be honored if they reviewed my submission.
