+++ vkCompViz/src/gpu.vulkan.cpp
@@ -785,7 +785,7 @@ void Vulkan::graphicsSubmit(size_t swapChainFrameID)
 
 void Vulkan::updateUniformBuffer(SwapChain::InFlight &inFlight)
 {
-    vmaCopyMemoryToAllocation(*inFlight.buffers.uniform->allocator, currentUniformBufferData.data(), inFlight.buffers.uniform->allocation, 0, currentUniformBufferData.size()*sizeof(uint32_t)); 
+    vmaCopyMemoryToAllocation(*inFlight.buffers.uniform->allocator, currentUniformBufferData.data(), inFlight.buffers.uniform->allocation, 0, currentUniformBufferData.size()*sizeof(uint32_t));
 }
 
 void Vulkan::updateShaderStorageBuffer(SwapChain::InFlight &inFlight)
@@ -960,7 +960,7 @@ vk::ImageCreateInfo &Vulkan::CreateInfo::image(vk::Format imageFormat, Resolutio
     auto sharingMode = vk::SharingMode::eConcurrent;
     queueFamilyIndices.clear();
     queueFamilyIndices.push_back(graphicsQueueID());
-    if (computeQueueID() != graphicsQueueID())
+    if(computeQueueID() != graphicsQueueID())
         queueFamilyIndices.push_back(computeQueueID());
     else
         sharingMode = vk::SharingMode::eExclusive;
@@ -972,17 +972,17 @@ vk::ImageCreateInfo &Vulkan::CreateInfo::image(vk::Format imageFormat, Resolutio
         usageFlags = vk::ImageUsageFlagBits::eDepthStencilAttachment;
 
     imageCreateInfo
-        .setFormat(imageFormat)
-        .setImageType(vk::ImageType::e2D)
-        .setExtent({resolution.width, resolution.height, 1})
-        .setMipLevels(1)
-        .setArrayLayers(1)
-        .setTiling(vk::ImageTiling::eOptimal)
-        .setInitialLayout(vk::ImageLayout::eUndefined)
-        .setUsage(usageFlags)
-        .setSharingMode(sharingMode)
-        .setQueueFamilyIndices(queueFamilyIndices)
-        .setSamples(vk::SampleCountFlagBits::e1);
+    .setFormat(imageFormat)
+    .setImageType(vk::ImageType::e2D)
+    .setExtent({resolution.width, resolution.height, 1})
+    .setMipLevels(1)
+    .setArrayLayers(1)
+    .setTiling(vk::ImageTiling::eOptimal)
+    .setInitialLayout(vk::ImageLayout::eUndefined)
+    .setUsage(usageFlags)
+    .setSharingMode(sharingMode)
+    .setQueueFamilyIndices(queueFamilyIndices)
+    .setSamples(vk::SampleCountFlagBits::e1);
     return imageCreateInfo;
 }
 
@@ -1355,7 +1355,7 @@ void Vulkan::createSwapChainFrames()
         inFlight.commandBuffers.graphics.emplace(std::move(graphicsCommandBuffers[i]));
         inFlight.commandBuffers.compute.emplace(std::move(computeCommandBuffers[i]));
         createInfo.createFrameSync(inFlight);
-        inFlight.buffers.uniform = memory.buffer(vk::BufferUsageFlagBits::eUniformBuffer, currentUniformBufferData.size()*sizeof(uint32_t));
+        inFlight.buffers.uniform = memory.buffer(vk::BufferUsageFlagBits::eUniformBuffer, currentUniformBufferData.size() * sizeof(uint32_t));
         swapChain.inFlight.back().descriptorSet.emplace(std::move(descriptorSets[i]));
         auto outputImages = createInfo.outputImages();
[…]
