+++ paper/paper.bib
@@ -1,182 +1,0 @@
-@inproceedings{plebanski2025,
-author = {Plebański, Piotr and Kelm, Anna and Hajder, Marcin},
-year = {2025},
-month = {01},
-pages = {111-119},
-title = {Efficiency and Development Effort of OpenCL Interoperability in Vulkan and OpenGL Environments: A Comparative Case Study},
-doi = {10.5220/0013529000003964}
-}
-
-@inproceedings{henriksen2024,
-author = {Henriksen, Troels},
-title = {A Comparison of OpenCL, CUDA, and HIP as Compilation Targets for a Functional Array Language},
-year = {2024},
-isbn = {9798400711008},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-doi = {10.1145/3677997.3678226},
-booktitle = {Proceedings of the 1st ACM SIGPLAN International Workshop on Functional Programming for Productivity and Performance},
-pages = {1–9},
-numpages = {9},
-keywords = {GPU, functional programming, parallel programming, performance measurement},
-location = {Milan, Italy},
-series = {FProPer 2024}
-}
-
-@inproceedings{lopez2025,
-author = {Mu\~{n}oz Lopez, Jose Emilio and Winston, Steven},
-title = {Hands-on Vulkan® Ray Tracing With Dynamic Rendering},
-year = {2025},
-isbn = {9798400715501},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-doi = {10.1145/3721251.3742861},
-abstract = {This Lab invites participants to explore how today’s Vulkan ecosystem tools and resources allow everyone to write real-time ray-traced effects with far less boilerplate than in the past. Building on a trimmed version of the open-source Khronos Vulkan Tutorial [Khronos 2025a] and the vk-bootstrap [Giessen 2025] and Vulkan-RAII [Khronos 2025c] helper libraries, we will incrementally assemble a renderer that combines rasterization with ray tracing to implement accurate shadows. The session balances bite-sized coding intervals with short concept discussions, so that every participant leaves with a runnable project and a mental map of where to dig deeper afterward.},
-booktitle = {Proceedings of the Special Interest Group on Computer Graphics and Interactive Techniques Conference Labs},
-articleno = {26},
-numpages = {1},
-location = {
-},
-series = {SIGGRAPH Labs '25}
-}
-
-@ARTICLE{rossant2021,
-  author={Rossant, Cyrille and Rougier, Nicolas P.},
-  journal={Computing in Science & Engineering}, 
-  title={High-Performance Interactive Scientific Visualization With Datoviz via the Vulkan Low-Level GPU API}, 
-  year={2021},
-  volume={23},
-  number={4},
-  pages={85-90},
-  doi={10.1109/MCSE.2021.3078345}}
-
-\
-@InProceedings{lavrivc2018,
-author="Lavri{\v{c}}, Primo{\v{z}}
-and Bohak, Ciril
-and Marolt, Matija",
-editor="De Paolis, Lucio Tommaso
-and Bourdot, Patrick",
-title="Vulkan Abstraction Layer for Large Data Remote Rendering System",
-booktitle="Augmented Reality, Virtual Reality, and Computer Graphics",
-year="2018",
-publisher="Springer International Publishing",
-address="Cham",
-pages="480--488",
-doi="10.1007/978-3-319-95270-3_40",
-isbn="978-3-319-95270-3"
-}
-
-@article{unter2023,
[…]
