### 2025-08-26 #8871 @editorialbot
[REVIEW]: vkCompViz: Universal C++ Library for GPU-Based Experiments
**Submitting author:** @ichlubna (<a href="http://orcid.org/0000-0003-3126-0545">Tomáš Chlubna</a>)
**Repository:** https://github.com/ichlubna/vkCompViz
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.5
