### 2025-09-15 afc829a
Argument parser check, more safe Slang reflection
