### 2025-09-15 8a902c0
Fixed default uniform limits
