### 2025-09-19 6a8517f
Extended documentation
