### 2025-09-25 #8871 @dhelmrich
@junghans I have concluded my review (checked all boxes) and the software uses a new cxx standard that needs to be paid attention to when building and compiling the library. But since most modern standards in C++ eventually get wide-spread implementation at some point I do not believe this is an issue. I can furthermore also reproduce the functinoality and data presented in the paper and assess that it is ready for publication, unless I missed something. I also did not see anything in the guidelines about what I am supposed to do after concluding my checklist when the review is generally still open - did I miss something?
