### 2025-09-25 #8871 @ichlubna
Thank you @dhelmrich for your review. Your remarks helped me a lot.
Thank you @boonth. Yes, as @junghans said, the new C++ standard might be a bit tricky. Please note the README, I use certain version of CMake because the modules support is still experimental. My goal is to have the library prepared for modern C++ which will be more widespread in a few years.
