### 2025-09-25 1843252
CMakeLists.txt: bump CMake Version to 4.1.1
