### 2025-09-25 952b80a
README.md: remove now obsolete section
