### 2025-09-25 #8871 @ichlubna
> Couldn't you require CMake-4.1.1 as I suggested in [ichlubna/vkCompViz#8](https://github.com/ichlubna/vkCompViz/pull/8)?

Hmm, that is quite a nice and cleaner solution, better than the README note, thank you. The only thing is that the proposal fixes minimal version but a higher CMake version might be problematic if the experimental hash changes. Anyway, I like this. I see that your PR is marked as WIP so I will wait for possible other changes before merging or implementing it.
