### 2025-09-25 8721dba
CMakeLists.txt: guard CMAKE_EXPERIMENTAL hash
