### 2025-10-06 #8871 @boonth
@junghans Sorry, I was able to compile on a Fedora machine, but I'm getting runtime errors when trying to run the examples. Compiling and running vulkan examples also produce runtime errors, so I think there is some issue with my environment. I will try to build and run on a Mac laptop, next.

@ichlubna I read the paper and it seems fine. There were some typos I found in the paper and documentation, listed below.

- Page 1, Statment of Need, second paragraph
  - libary -> library
- Figure 3 caption
  - masurements -> measurements
- vkCompViz/examples/simpleBlending/README.md, vkCompViz/examples/3DViewer/README.md, vkCompViz/examples/parallelReduction/README.md
  - buikd -> build
