+++ examples/parallelReduction/README.md
@@ -2,7 +2,7 @@
 This program computes a sum of a random array of float values and compares the CPU and GPU versions. The GPU version uses shared memory and parallel reduction scheme.
 
 ## Example usage
-[Build](../../README.md) the examples. Then run in the `buikd` directory:
+[Build](../../README.md) the examples. Then run in the `build` directory:
 
     cd examples/parallelReduction/; ./parallelReduction -s 1000000; cd -
 
