+++ examples/3DViewer/README.md
@@ -2,7 +2,7 @@
 This program can open an OBJ file and render it in a window. The OBJ file must contain vertices and normals and the faces are expected to be triangles, not quads.
 
 ## Example usage
-[Build](../../README.md) the examples. Then run in the `buikd` directory:
+[Build](../../README.md) the examples. Then run in the `build` directory:
 
     cd examples/3DViewer/; ./3DViewer ../../../examples/3DViewer/data/test.obj; cd -
 
