### 2025-10-06 #8871 @junghans
@boonth this review issue might be not the right place to discuss those runtime errors in detail, but could you open an issue here: https://github.com/ichlubna/vkCompViz/issues?

@ichlubna could you work with @boonth to document these runtime errors in the README?
