### 2025-10-06 ichlubna/vkCompViz#9 @boonth
Runtime error with simpleBlending
I got the code to compile, and when running the simpleBlending example, I got the following runtime error:

```
yod@grace:~/tmp/joss_review/vkCompViz/build/examples/simpleBlending$ ./simpleBlending -i1 ../../../examples/simpleBlending/data/A.webp -i2 ../../../examples/simpleBlending/data/B.webp -f 0.5 -o ../../result.webp
[in @ 0x1b72b2c0] Changing video frame properties on the fly is not supported by all filters.
[in @ 0x1b72b2c0] filter context - w: 800 h: 534 fmt: 0 csp: unknown range: unknown, incoming frame - w: 800 h: 534 fmt: 0 csp: bt470bg range: tv pts_time: 0
[in @ 0x1b72bbc0] Changing video frame properties on the fly is not supported by all filters.
[in @ 0x1b72bbc0] filter context - w: 800 h: 534 fmt: 0 csp: unknown range: unknown, incoming frame - w: 800 h: 534 fmt: 0 csp: bt470bg range: tv pts_time: 0
[in @ 0x1ba63100] Changing video frame properties on the fly is not supported by all filters.
[in @ 0x1ba63100] filter context - w: 320 h: 240 fmt: 0 csp: unknown range: unknown, incoming frame - w: 320 h: 240 fmt: 0 csp: bt470bg range: tv pts_time: 0
MESA-INTEL: warning: Haswell Vulkan support is incomplete
The following devices are available:
Name: AMD Radeon RX 6700 XT (RADV NAVI22)
- Capabilities: Compute Discrete Graphics Present Required-extensions Swapchain
- ID: 17719716715334191207243862447183106121113131
Name: Intel(R) HD Graphics 4600 (HSW GT2)
- Capabilities: Compute Graphics Present
- ID: 2335117121644219189448746219991688148142
Name: llvmpipe (LLVM 20.1.8, 256 bits)
- Capabilities: Compute Graphics Present Required-extensions Swapchain
- ID: 50534649465797979797979797979797

Selected as the most suitable:
Name: AMD Radeon RX 6700 XT (RADV NAVI22)
- Capabilities: Compute Discrete Graphics Present Required-extensions Swapchain
- ID: 17719716715334191207243862447183106121113131

PhysicalDevice::createDevice: ErrorFeatureNotPresent
```
