### 2025-10-06 #8871 @boonth
Made an issue about the runtime error here: https://github.com/ichlubna/vkCompViz/issues/9. As I said earlier, I suspect it is more about my environment/drivers than anything with the code.
