### 2025-10-07 2db29c3
Features check
