+++ README.md
@@ -31,6 +31,12 @@ A path to Artistic Style (astyle) program can be defined with `ccmake` and the p
 ```
 ninja format
 ```
+
+The library requires certain modern GPU features. 
+Please read the output messages when running the examples. 
+The library tries to automatically select the best available device, which supports most of the requirements. 
+However, lack of some features might block the examples from running.
+
 ## Usage
 See the existing examples for reference. Especially, [Simple Blending](examples/simpleBlending), [Parallel Reduction](examples/parallelReduction), and [3D Viewer](examples/3DViewer) are simple and contain explanatory comments. It is recommended to use them as templates for new examples. The API of the library is described in the [documentation](https://ichlubna.github.io/vkCompViz/classvkCompViz_1_1App.html).
 
