### 2025-10-09 8887634
Logo and typo fix
