### 2025-10-09 ichlubna/vkCompViz#10 @boonth
Compiling on Mac
I tried compiling on my Mac laptop (Apple M3 Max). After installing the dependencies, I got the following error:

```
CMake Error at examples/3DViewer/CMakeLists.txt:19 (target_compile_features):
  target_compile_features The compiler feature "cxx_std_26" is not known to
  CXX compiler

  "AppleClang"

  version 17.0.0.17000319.
```

So I installed llvm with Homebrew. Using Homebrew's Clang, I was able to get CMake to configure, but when generating files, the following error occurred:

```
-- Configuring done (3.0s)
CMake Error in vkCompViz/CMakeLists.txt:
  The "CXX_MODULE_STD" property on the target "VulkanHppModule" requires that
  the "__CMAKE::CXX26" target exist, but it was not provided by the
  toolchain.  Reason:

    `libc++.modules.json` resource does not exist
```

Is it correct that AppleClang does not support cxx_std_26? I'm not sure how to check.

I have not been able to move past the error with not finding the json file.

Related to https://github.com/openjournals/joss-reviews/issues/8871
