### 2025-12-02 #8871 @boonth
> Ping [@boonth](https://github.com/boonth)

@junghans I will get back to trying to compile to code on my Mac this week. If that doesn't work, I am planning on trying to use a docker image of ubuntu and use the script in the repository to try to compile the code.
