### 2026-01-13 b3a0330
Build script fixed and docker instructions added
