+++ scripts/buildOnUbuntu-24_04.sh
@@ -1,33 +1,36 @@
 #!/bin/bash
 set -e
-# If used in docker:
-# docker run --privileged -it --name vkCompViz ubuntu:24.04 bash
 
 # When used in Docker as root
-if ! command -v sudo &> /dev/null; then
-	apt update
-    apt install -y sudo
+if [ "$1" == "docker" ]; then
+    if [ -d "vkCompViz" ]; then
+        echo "The directory vkCompViz already exists."
+        exit 0
+    fi
+    if ! command -v sudo &> /dev/null; then
+        apt update
+        apt install -y sudo
+    fi
 fi
 # Installing basic dev tools
+export DEBIAN_FRONTEND=noninteractive
+sudo ln -fs /usr/share/zoneinfo/Europe/Prague /etc/localtime
+echo "Europe/Prague" | sudo tee /etc/timezone
 sudo apt update
 sudo apt install software-properties-common -y
 sudo add-apt-repository ppa:ubuntuhandbook1/ffmpeg7 -y; sudo apt update
-sudo apt install sudo ninja-build git wget tzdata libglfw3-dev ffmpeg libavfilter-dev libavformat-dev libavcodec-dev libavutil-dev libswscale-dev libavdevice-dev build-essential -y
+sudo apt install sudo ninja-build git wget gnupg software-properties-common tzdata libglfw3-dev ffmpeg libavfilter-dev libavformat-dev libavcodec-dev libavutil-dev libswscale-dev libavdevice-dev build-essential -y
 
 # Installing Clang 20
-sudo apt update
-export DEBIAN_FRONTEND=noninteractive
-sudo ln -fs /usr/share/zoneinfo/Europe/Prague /etc/localtime
-sudo dpkg-reconfigure --frontend noninteractive tzdata
-sudo apt install -y wget software-properties-common gnupg
 wget -qO - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo gpg --dearmor > /usr/share/keyrings/llvm-archive-keyring.gpg
-echo "deb [signed-by=/usr/share/keyrings/llvm-archive-keyring.gpg] http://apt.llvm.org/noble/ llvm-toolchain-noble main" | sudo tee /etc/apt/sources.list.d/llvm.list
 echo "deb [signed-by=/usr/share/keyrings/llvm-archive-keyring.gpg] http://apt.llvm.org/noble/ llvm-toolchain-noble-20 main" | sudo tee -a /etc/apt/sources.list.d/llvm.list
 sudo apt update; sudo apt install -y clang-20 libc++-20-dev libc++abi-20-dev
 
 # Cloning the project in Docker
-#git clone --recursive https://github.com/ichlubna/vkCompViz.git
-#cd vkCompViz/
+if [ "$1" == "docker" ]; then
+    git clone --recursive https://github.com/ichlubna/vkCompViz.git
+    cd vkCompViz/
+fi
 
 # Preparing submodules and build dir
 git submodule update --init --recursive
@@ -65,7 +68,3 @@ sudo apt update; sudo apt install vulkan-sdk -y
 cd ..
[…]
