+++ scripts/README.md
@@ -7,7 +7,7 @@ Make sure the Docker daemon is running:
 
     sudo systemctl start docker
 
-The following commands will download the script, run it in a new Ubuntu Docker machine and stay in the machine so the results can be checked. Note that the script uses a parameter `docker` to indicate it is not run in the GitHub Workflow.
+The following commands will download the script, run it in a new Ubuntu Docker machine and stay in the machine so the results can be checked. Note that the script uses a parameter `docker` to indicate it is not run in the GitHub Workflow. The image used for the testing is `ubuntu:24.04`.
 
     wget https://raw.githubusercontent.com/ichlubna/vkCompViz/refs/heads/main/scripts/buildOnUbuntu-24_04.sh
     docker run --privileged -it --name vkCompViz -v "$(pwd)/buildOnUbuntu-24_04.sh:/buildOnUbuntu-24_04.sh:ro" ubuntu:24.04 bash -c "bash /buildOnUbuntu-24_04.sh docker; exec bash"
