### 2026-01-15 #8871 @junghans
@ichlubna, can you take care of the author item above and decide if you want to merge https://github.com/ichlubna/vkCompViz/pull/8 before the doing version tag.
