### 2026-01-15 6e9f810
Compile script changed and vulkan API changes reflected
