### 2026-01-15 #8871 @ichlubna
@junghans here is the rest of the tasks:
The latest version including the changes from the review is v1.0.5.
Zenodo DOI is https://doi.org/10.5281/zenodo.18258201
