### 2026-01-16 7536012
Update paper.bib
