+++ paper/paper.bib
@@ -132,6 +132,7 @@ address = {New York, NY, USA},
 volume = {48},
 number = {4},
 issn = {0360-0300},
+url = {https://doi.org/10.1145/2873053},
 doi = {10.1145/2873053},
 month = mar,
 articleno = {60},
@@ -179,3 +180,23 @@ volume = {23},
 journal = {Network (Bristol, England)},
 doi = {10.3109/0954898X.2012.739292}
 }
+
+@article{trott2022,
+  author={Trott, Christian R. and Lebrun-Grandié, Damien and Arndt, Daniel and Ciesko, Jan and Dang, Vinh and Ellingwood, Nathan and Gayatri, Rahulkumar and Harvey, Evan and Hollman, Daisy S. and Ibanez, Dan and Liber, Nevin and Madsen, Jonathan and Miles, Jeff and Poliakoff, David and Powell, Amy and Rajamanickam, Sivasankaran and Simberg, Mikael and Sunderland, Dan and Turcksin, Bruno and Wilke, Jeremiah},
+  journal={IEEE Transactions on Parallel and Distributed Systems},
+  title={Kokkos 3: Programming Model Extensions for the Exascale Era},
+  year={2022},
+  volume={33},
+  number={4},
+  pages={805-817},
+  doi={10.1109/TPDS.2021.3097283}}
+
+@INPROCEEDINGS{beckingsale2019,
+  author={Beckingsale, David A. and Burmark, Jason and Hornung, Rich and Jones, Holger and Killian, William and Kunen, Adam J. and Pearce, Olga and Robinson, Peter and Ryujin, Brian S. and Scogland, Thomas RW},
+  booktitle={2019 IEEE/ACM International Workshop on Performance, Portability and Productivity in HPC (P3HPC)}, 
+  title={RAJA: Portable Performance for Large-Scale Scientific Applications}, 
+  year={2019},
+  volume={},
+  number={},
+  pages={71-81},
+  doi={10.1109/P3HPC49587.2019.00012}}
