### 2026-01-16 #8871 @ichlubna
@junghans I have shortened the statement of need as:

Removed as a redundant sentence, instead going right to the description of the problems:
_However, using these APIs presents various problems._

This sentence repeats the previously stated advantages so it is removed now:
The goal is to reduce the time required for GPU programming, particularly since numerous scientific experiments may not produce significant results, and it is inefficient to spend excessive time learning about existing APIs.

Shortened the following sentences:
_OpenCL is a multiplatform and vendor-free API that contains a lot of features, but, in practice, GPU drivers usually do not implement its entire specification._
to:
_OpenCL is a vendor-free multiplatform API with a lot of features that are usually not fully supported by drivers._

_The APIs referenced earlier are tailored only for GPGPU computations and lack direct access to GPU rendering features._
to:
_These APIs are only for GPGPU computations and lack access to GPU rendering features._

_In contrast, subsequent APIs are capable of using rendering pipelines and GPGPU functions._
to:
_The subsequent APIs can use rendering pipelines and GPGPU functions._

_The main disadvantage of Vulkan is its verbosity, which makes it difficult to use for quick experiments, as a substantial portion of the code is necessary even for basic functionality._
to:
The verbosity of Vulkan makes it difficult to use for quick experiments. A substantial portion of the code is necessary even for basic functionality.

_With this library, researchers do not need to learn about complex GPU-related APIs and can quickly conduct scientific experiments on GPU.
This tackles a frequent issue in the scientific domain, where bringing an experimental concept to life usually requires a considerable amount of time owing to the need to acquire novel technological skills and address numerous small, implementation-related challenges._
to:
_Users do not need to study complex GPU APIs to conduct scientific experiments.
This addresses a frequent issue in science, where implementing an experimental concept often requires a considerable amount of time due to technical difficulties._

Added this comparison:
_Kokkos [@trott2022] and RAJA [@beckingsale2019] are C++ abstractions for performance-portable parallel computation across CPUs and GPUs.
In contrast, vkCompViz is GPU-oriented.
Its goal is not only to simplify parallel computing but also visualization and data loading and provide access to additional GPU capabilities, such as rendering and other features exposed by Vulkan._
