### 2026-01-16 #8871 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
