### 2026-01-16 4f4b257
minor changes in paper.md
