### 2026-01-17 #8871 @ichlubna
Thank you @danielskatz, all the suggestions look good. I have merged them.
As some researchers mentioned in openjournals/joss#1283, it is problematic to assign JOSS papers to research projects due to JOSS absence in WoS or Scopus databases. Therefore, I have no acknowledgements to write.
