### 2026-01-17 #8871 @danielskatz
👋 @ichlubna - I think you are misunderstanding the purpose of acknowledgments, which are to let the reader know about any sources of funding or other support that you think are important to the work. This has nothing to do with where the work is indexed.  However, it's up to you to choose, so I'll proceed to generate a final proof now.
