### 2026-01-17 #8871 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
