### 2026-01-17 #8871 @ichlubna
@danielskatz Thank you. Yea, it's probably about institutional conventions. We usually write the research projects statements in the acknowledgements that were behind the given paper. They often expect indexed results.
Anyway thank you all @danielskatz, @junghans, @boonth and @dhelmrich for your assistance. This open review process was a great experience!
