### 2026-01-17 40f6492
JOSS citation
