### 2026-01-17 bfbf25e
JOSS citation
