+++ CITATION.cff
@@ -1,12 +1,26 @@
-@article{Chlubna2026,
-title = {vkCompViz: Universal C++ Library for GPU-Based Experiments},
-author = {Chlubna, Tomas},
-year = 2026,
-journal = {Journal of Open Source Software},
-publisher = {The Open Journal},
-volume = 11,
-number = 117,
-pages = 8871,
-doi = {10.21105/joss.08871},
-url = {https://doi.org/10.21105/joss.08871}
-}
+cff-version: "1.2.0"
+authors:
+- family-names: Chlubna
+  given-names: Tomas
+  orcid: "https://orcid.org/0000-0003-3126-0545"
+doi: 10.5281/zenodo.18258201
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Chlubna
+    given-names: Tomas
+    orcid: "https://orcid.org/0000-0003-3126-0545"
+  date-published: 2026-01-17
+  doi: 10.21105/joss.08871
+  issn: 2475-9066
+  issue: 117
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 8871
+  title: "vkCompViz: Universal C++ Library for GPU-Based Experiments"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.08871"
+  volume: 11
+title: "vkCompViz: Universal C++ Library for GPU-Based Experiments"
