### 2026-01-17 #8871 @danielskatz
Congratulations to @ichlubna (Tomáš Chlubna) on your JOSS publication!!

JOSS completely depends on volunteers; we couldn't do this without you
